Migrate to jest for testing
