+++ __tests__/fixture/comment-with-trailing-space.scss
