+++ __tests__/fixture/simple-comment.scss
