+++ __tests__/fixture/single-quoted-import-path.scss
