+++ __tests__/fixture/single-quoted-import.scss
