+++ package.json
@@ -4,7 +4,7 @@
   "description": "A tokenzier for Sass' SCSS syntax",
   "main": "index.js",
   "scripts": {
-    "test": "mocha",
+    "test": "jest",
     "clean": "rm lib/*",
     "build": "npm run clean; babel src/ --out-dir lib"
   },
@@ -35,10 +35,11 @@
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-jest": "^22.4.3",
     "babel-preset-es2015": "^6.24.1",
-    "chai": "^4.1.2",
     "glob": "^7.1.2",
-    "mocha": "^5.0.1",
+    "jest": "^22.4.3",
     "sass-spec": "^3.5.1"
   }
 }
